+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Report e-mail in HTML format
 - Option to ignore backup parts from files that didn't change in disk
 - Download many archives in parallel
+- Using context from AWS library for user cancellation
 
 ### Fixed
 - Ignore unmodified files when choosing the backup parts to download